Move query_apl to top-level query
