+++ src/datasets/mod.rs
@@ -20,7 +20,7 @@
 //!       }),
 //!     ]).await?;
 //!
-//!     let res = client.datasets.apl_query("['my-dataset'] | count", None).await?;
+//!     let res = client.query("['my-dataset'] | count", None).await?;
 //!     assert_eq!(1, res.status.rows_matched);
 //!
 //!     client.datasets.delete("my-dataset").await?;
